+++ Mathematics/Vector2.cs
@@ -4,21 +4,21 @@ namespace Mathematics;
 
 public readonly struct Vector2
 {
-	private readonly float _x, _y;
+	private readonly decimal _x, _y;
 
 	public Vector2()
 	{
-		_x = 0.0F;
-		_y = 0.0F;
+		_x = 0.0M;
+		_y = 0.0M;
 	}
-	public Vector2(float x, float y)
+	public Vector2(decimal x, decimal y)
 	{
 		_x = x;
 		_y = y;
 	}
 
-	public float X => _x;
-	public float Y => _y;
+	public decimal X => _x;
+	public decimal Y => _y;
 
 	public static Vector2 Zero => new Vector2(0, 0);
